refactor: promote scheduler protocol kernel
