+++ src/domain/mod.rs
@@ -1,5 +1,6 @@
 //! Canonical runtime domain records.
 
+pub mod scheduler_protocol;
 pub mod work_item;
 pub mod workspace;
